Add shortcut to spawn new terminal in the current dir
Build on bakkeby's doule fork version. Use /proc/self/exe as the binary,
and argv[0] as the called name. This allows to have several different
st's and running from build tree etc. Also skip the unnecessary readlink
part, shell shall do that anyway. Futher mark the pty master as cloexec,
so the fd's are not leaked to spawned processes.